refactor and simplify DRS bundle expansion
